moved examples into the client so i can install it easily
